Inline createActors into actor_test
It is the only place where this is used so it might as well be
a private function.
